\name{erAlpha}
\alias{erAlpha}
\docType{data}
\title{
GSE14664 ERalpha bining in MCF-7 cell line
}
\description{
ChIP-Seq was used to map ERalpha binding sites and to profile changes in RNA polymerase II (RNAPII) occupancy in MCF-7 cells in response to estradiol (E2), tamoxifen or fulvestrant
}
\usage{data(erAlpha)}
\format{
  The format is:
List of 5
 $ pdata    : chr [1:4, 1:2] "GSM365925" "GSM365926" "GSM365927" "GSM365928" ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:4] "GSM365925" "GSM365926" "GSM365927" "GSM365928"
  .. ..$ : chr [1:2] "title" "des"
 $ GSM365925:'data.frame':	240 obs. of  5 variables:
  ..$ CHROMOSOME: chr [1:240] "chr1" "chr1" "chr1" "chr1" ...
  ..$ POSITION  : int [1:240] 557137 559461 9932982 17436548 43230435 91625079 107670920 115101584 121185889 121186256 ...
  ..$ END       : int [1:240] 557914 560279 9933750 17436933 43231478 91626093 107671824 115102160 121186254 121186532 ...
  ..$ LENGTH    : int [1:240] 777 818 768 385 1043 1014 904 576 365 276 ...
  ..$ SCORE     : num [1:240] 10.97 24.96 9.79 10.31 24.17 ...
 $ GSM365926:'data.frame':	10205 obs. of  5 variables:
  ..$ CHROMOSOME: chr [1:10205] "chr1" "chr1" "chr1" "chr1" ...
  ..$ POSITION  : int [1:10205] 557112 559459 998864 999399 1004343 1070346 1305474 1358287 1776987 1820476 ...
  ..$ END       : int [1:10205] 558114 560286 999397 999604 1005146 1071080 1306502 1358744 1777750 1821168 ...
  ..$ LENGTH    : int [1:10205] 1002 827 533 205 803 734 1028 457 763 692 ...
  ..$ SCORE     : num [1:10205] 29 32 137.5 15 35.2 ...
 $ GSM365927:'data.frame':	8854 obs. of  5 variables:
  ..$ CHROMOSOME: chr [1:8854] "chr1" "chr1" "chr1" "chr1" ...
  ..$ POSITION  : int [1:8854] 557112 559461 998902 999401 999502 1004441 1358375 1774793 1777038 1922805 ...
  ..$ END       : int [1:8854] 557932 560281 999399 999500 999898 1005121 1358834 1775579 1777738 1923836 ...
  ..$ LENGTH    : int [1:8854] 820 820 497 99 396 680 459 786 700 1031 ...
  ..$ SCORE     : num [1:8854] 28.9 26.9 119.6 14.6 13.5 ...
 $ GSM365928:'data.frame':	4284 obs. of  5 variables:
  ..$ CHROMOSOME: chr [1:4284] "chr1" "chr1" "chr1" "chr1" ...
  ..$ POSITION  : int [1:4284] 557140 559649 998957 1004390 3244680 6465563 6594411 6942285 7430488 8373888 ...
  ..$ END       : int [1:4284] 557846 560288 999352 1005025 3245330 6466221 6594758 6942720 7430972 8374584 ...
  ..$ LENGTH    : int [1:4284] 706 639 395 635 650 658 347 435 484 696 ...
  ..$ SCORE     : num [1:4284] 31.9 30.9 63.8 12.9 27.4 ...
}
\details{
Welboren WJ, van Driel MA, Janssen-Megens EM, van Heeringen SJ et al. ChIP-Seq of ERalpha and RNA polymerase II defines genes differentially responding to ligands. EMBO J 2009 May 20;28(10):1418-28.
}
\source{
http://www.eh3.uc.edu/GenomicsPortals/analyzeExp_tl.jsp?exp=GSE14664ERalphaPeaks&db=Hg18Refseq
}
\references{
Welboren WJ, van Driel MA, Janssen-Megens EM, van Heeringen SJ et al. ChIP-Seq of ERalpha and RNA polymerase II defines genes differentially responding to ligands. EMBO J 2009 May 20;28(10):1418-28.
}
\examples{
data(erAlpha)
## maybe str(erAlpha) ; plot(erAlpha) ...
}
\keyword{datasets}
