% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group_gene_mean.R
\name{get_group_gene_mean}
\alias{get_group_gene_mean}
\title{Get the Mean logcount for each Gene in each Group}
\usage{
get_group_gene_mean(sce, group_col = "cellType")
}
\arguments{
\item{sce}{SingleCellExperiment Object with logcount assay}

\item{group_col}{name of the column in the colData of sce that defines the group of interest}
}
\value{

}
\description{
Get the Mean logcount for each Gene in each Group
}
\examples{

get_group_gene_mean(sce_zero_test, group_col = "group")

}
