% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group_gene_rank.R
\name{get_group_gene_rank}
\alias{get_group_gene_rank}
\title{Get the Rank of the Mean logcount for each Gene in each Group}
\usage{
get_group_gene_rank(sce, group_col = "cellType")
}
\arguments{
\item{sce}{SingleCellExperiment Object with logcount assay}

\item{group_col}{name of the column in the colData of sce that defines the group of interest}
}
\value{
Named list of ranks for each gene
}
\description{
Get the Rank of the Mean logcount for each Gene in each Group
}
\examples{

get_group_gene_rank(sce_zero_test, group_col = "group")

}
