% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rank_invariance.R
\name{get_rank_invariance}
\alias{get_rank_invariance}
\title{Calculate the Rank Invariance of Each Gene from Cell and Group Ranks}
\usage{
get_rank_invariance(group_rank, cell_rank)
}
\arguments{
\item{cell_rank}{}
}
\value{
rank of invariance for each gene
}
\description{
Calculate the Rank Invariance of Each Gene from Cell and Group Ranks
}
\examples{
group_rank_test <- get_group_gene_rank(sce_zero_test, group_col = "group")
cell_rank_test <- rank_cells(sce_zero_test, group_col = "group")
rank_invar_test <- get_rank_invariance(group_rank_test, cell_rank_test)
}
