% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rank_invariance.R
\name{get_rank_invariance_express}
\alias{get_rank_invariance_express}
\title{Calculate the Rank Invariance of Each Gene from SCE object w/ groups}
\usage{
get_rank_invariance_express(sce, group_col = "cellType")
}
\arguments{
\item{sce}{SingleCellExperiment Object with logcount assay}

\item{group_col}{name of the column in the colData of sce that defines the group of interest}
}
\value{
rank of invariance for each gene
}
\description{
Calculate the Rank Invariance of Each Gene from SCE object w/ groups
}
\examples{
get_rank_invariance_express(sce_zero_test, group_col = "group")
}
