\name{mesFactors}
\alias{mesFactors}
\docType{data}
\title{
GSE11431 TF binding in MES
}
\description{
Mapping of binding sites for thirteen sequence specific transcription factors (Nanog, Oct4, STAT3, Smad1, Sox2, Zfx, c-Myc, n-Myc, Klf4, Esrrb, Tcfcp2l1, E2f1 and CTCF) and two transcription regulators (p300 and Suz12) in mouse embryonic stem cells.
}
\usage{data(mesFactors)}
\format{
  The format is:
List of 16
 $ pdata    : chr [1:15, 1:2] "GSM288345" "GSM288346" "GSM288347" "GSM288348" ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : Named chr [1:15] "GSM288345" "GSM288346" "GSM288347" "GSM288348" ...
  .. .. ..- attr(*, "names")= chr [1:15] "c..GSM288345....Nanog.txt.." "c..GSM288346....Oct4.txt.." "c..GSM288347....Sox2.txt.." "c..GSM288348....Smad1.txt.." ...
  .. ..$ : chr [1:2] "sample" "TranscriptionFactor"
 $ GSM288345:'data.frame':	10343 obs. of  5 variables:
  ..$ CHROMOSOME: chr [1:10343] "chr1" "chr1" "chr1" "chr1" ...
  ..$ POSITION  : int [1:10343] 3053032 3333837 3334422 3473143 3671806 3937230 3985018 4141099 4809772 4845709 ...
  ..$ END       : int [1:10343] 3053034 3333843 3334449 3473144 3671822 3937239 3985079 4141105 4809775 4845710 ...
  ..$ SCORE     : num [1:10343] 11.47 11.47 5.52 9.77 6.37 ...
  ..$ LENGTH    : num [1:10343] 3 7 28 2 17 10 62 7 4 2 ...
 $ GSM288346:'data.frame':	3761 obs. of  5 variables:
  ..$ CHROMOSOME: chr [1:3761] "chr1" "chr1" "chr1" "chr1" ...
  ..$ POSITION  : int [1:3761] 3671785 6438973 6448207 7037533 7441814 7722807 9428371 9894478 10218301 11263254 ...
  ..$ END       : int [1:3761] 3671827 6439000 6448226 7037540 7441815 7722839 9428390 9894498 10218303 11263261 ...
  ..$ SCORE     : num [1:3761] 6.56 7.29 8.74 10.93 7.29 ...
  ..$ LENGTH    : num [1:3761] 43 28 20 8 2 33 20 21 3 8 ...
 $ GSM288347:'data.frame':	4526 obs. of  5 variables:
  ..$ CHROMOSOME: chr [1:4526] "chr1" "chr1" "chr1" "chr1" ...
  ..$ POSITION  : int [1:4526] 3053046 3904283 3984948 4141020 4809647 4845623 6049391 6107823 6439093 6448166 ...
  ..$ END       : int [1:4526] 3053052 3904333 3984950 4141042 4809699 4845626 6049436 6107833 6439094 6448169 ...
  ..$ SCORE     : num [1:4526] 5.57 7.42 7.42 13.36 6.68 ...
  ..$ LENGTH    : num [1:4526] 7 51 3 23 53 4 46 11 2 4 ...
 $ GSM288348:'data.frame':	1126 obs. of  5 variables:
  ..$ CHROMOSOME: chr [1:1126] "chr1" "chr1" "chr1" "chr1" ...
  ..$ POSITION  : int [1:1126] 3334335 3479676 4845666 6049369 6448201 7037388 7116065 7722858 9428337 11263312 ...
  ..$ END       : int [1:1126] 3334449 3479748 4845688 6049430 6448212 7037513 7116070 7722906 9428353 11263319 ...
  ..$ SCORE     : num [1:1126] 7.5 7.5 6.43 8.57 8.57 ...
  ..$ LENGTH    : num [1:1126] 115 73 23 62 12 126 6 49 17 8 ...
 $ GSM288349:'data.frame':	20699 obs. of  5 variables:
  ..$ CHROMOSOME: chr [1:20699] "chr1" "chr1" "chr1" "chr1" ...
  ..$ POSITION  : int [1:20699] 4775733 4791635 4797713 4842239 4845997 4847638 4848448 4960808 5073213 5250860 ...
  ..$ END       : int [1:20699] 4775734 4791645 4797717 4842239 4846040 4847638 4848451 4960818 5073221 5250895 ...
  ..$ SCORE     : num [1:20699] 40.38 5.94 17.22 18.41 7.72 ...
  ..$ LENGTH    : num [1:20699] 2 11 5 1 44 1 4 11 9 36 ...
 $ GSM288350:'data.frame':	26910 obs. of  5 variables:
  ..$ CHROMOSOME: chr [1:26910] "chr1" "chr1" "chr1" "chr1" ...
  ..$ POSITION  : int [1:26910] 4049605 4406940 4483876 4513052 4604250 4651053 4681952 4775715 4792790 4800480 ...
  ..$ END       : int [1:26910] 4049669 4406954 4483876 4513076 4604259 4651065 4681998 4775717 4792793 4800481 ...
  ..$ SCORE     : num [1:26910] 5.69 13.66 74.54 7.4 6.54 ...
  ..$ LENGTH    : num [1:26910] 65 15 1 25 10 13 47 3 4 2 ...
 $ GSM288351:'data.frame':	39609 obs. of  5 variables:
  ..$ CHROMOSOME: chr [1:39609] "chr1" "chr1" "chr1" "chr1" ...
  ..$ POSITION  : int [1:39609] 3002834 4132776 4322743 4336458 4407000 4506825 4758646 4760140 4770383 4845841 ...
  ..$ END       : int [1:39609] 3002851 4132783 4322748 4336496 4407004 4506825 4758646 4760163 4770394 4845855 ...
  ..$ SCORE     : num [1:39609] 5.34 32.04 24.27 15.05 33.01 ...
  ..$ LENGTH    : num [1:39609] 18 8 6 39 5 1 1 24 12 15 ...
 $ GSM288352:'data.frame':	10338 obs. of  5 variables:
  ..$ CHROMOSOME: chr [1:10338] "chr1" "chr1" "chr1" "chr1" ...
  ..$ POSITION  : int [1:10338] 4797100 4797987 4847954 5009381 5073137 6204657 8571686 9454924 9530746 9634685 ...
  ..$ END       : int [1:10338] 4797130 4798004 4847973 5009398 5073161 6204667 8571686 9454958 9530747 9634686 ...
  ..$ SCORE     : num [1:10338] 19.55 9.31 30.72 9.31 16.75 ...
  ..$ LENGTH    : num [1:10338] 31 18 20 18 25 11 1 35 2 2 ...
 $ GSM288353:'data.frame':	2546 obs. of  5 variables:
  ..$ CHROMOSOME: chr [1:2546] "chr1" "chr1" "chr1" "chr1" ...
  ..$ POSITION  : int [1:2546] 7116034 7299074 9531064 10069835 11263349 12729936 13070426 13079926 14298026 14406473 ...
  ..$ END       : int [1:2546] 7116041 7299159 9531067 10069859 11263353 12729939 13070429 13079957 14298048 14406500 ...
  ..$ SCORE     : num [1:2546] 6.69 6.02 6.02 11.37 12.04 ...
  ..$ LENGTH    : num [1:2546] 8 86 4 25 5 4 4 32 23 28 ...
 $ GSM288354:'data.frame':	10875 obs. of  5 variables:
  ..$ CHROMOSOME: chr [1:10875] "chr1" "chr1" "chr1" "chr1" ...
  ..$ POSITION  : int [1:10875] 4482767 4775858 4792828 4847633 5127270 5218163 7079100 7116068 7196237 7388173 ...
  ..$ END       : int [1:10875] 4482770 4775907 4792840 4847669 5127316 5218179 7079110 7116070 7196241 7388189 ...
  ..$ SCORE     : num [1:10875] 8.46 5.64 10.34 7.52 9.4 ...
  ..$ LENGTH    : num [1:10875] 4 50 13 37 47 17 11 3 5 17 ...
 $ GSM288355:'data.frame':	21647 obs. of  5 variables:
  ..$ CHROMOSOME: chr [1:21647] "chr1" "chr1" "chr1" "chr1" ...
  ..$ POSITION  : int [1:21647] 4482704 4483406 4581238 4651025 4775732 4792776 4797741 4800507 6204537 6439182 ...
  ..$ END       : int [1:21647] 4482715 4483411 4581259 4651053 4775735 4792840 4797743 4800545 6204542 6439191 ...
  ..$ SCORE     : num [1:21647] 22.8 14.87 6.44 15.86 41.63 ...
  ..$ LENGTH    : num [1:21647] 12 6 22 29 4 65 3 39 6 10 ...
 $ GSM288356:'data.frame':	3422 obs. of  5 variables:
  ..$ CHROMOSOME: chr [1:3422] "chr1" "chr1" "chr1" "chr1" ...
  ..$ POSITION  : int [1:3422] 4842133 5073201 9530642 9929283 11443723 13645615 16089445 20805381 30893750 30894463 ...
  ..$ END       : int [1:3422] 4842148 5073215 9530652 9929293 11443733 13645615 16089449 20805381 30893790 30894487 ...
  ..$ SCORE     : num [1:3422] 11.32 8.63 19.95 22.64 5.39 ...
  ..$ LENGTH    : num [1:3422] 16 15 11 11 11 1 5 1 41 25 ...
 $ GSM288357:'data.frame':	7182 obs. of  5 variables:
  ..$ CHROMOSOME: chr [1:7182] "chr1" "chr1" "chr1" "chr1" ...
  ..$ POSITION  : int [1:7182] 4775791 4797915 4842056 5073102 6204811 7114794 7387792 9530636 9533296 9685381 ...
  ..$ END       : int [1:7182] 4775811 4797916 4842057 5073107 6204819 7114809 7387809 9530637 9533303 9685392 ...
  ..$ SCORE     : num [1:7182] 11.13 10.39 13.35 6.68 7.42 ...
  ..$ LENGTH    : num [1:7182] 21 2 2 6 9 16 18 2 8 12 ...
 $ GSM288359:'data.frame':	524 obs. of  5 variables:
  ..$ CHROMOSOME: chr [1:524] "chr1" "chr1" "chr1" "chr1" ...
  ..$ POSITION  : int [1:524] 12677639 12729916 13070392 26632128 29445300 36930531 42003241 59367837 71069233 72195295 ...
  ..$ END       : int [1:524] 12677655 12729921 13070394 26632132 29445322 36930538 42003260 59367862 71069237 72195340 ...
  ..$ SCORE     : num [1:524] 5.28 9.9 6.6 5.28 5.28 ...
  ..$ LENGTH    : num [1:524] 17 6 3 5 23 8 20 26 5 46 ...
 $ GSM288360:'data.frame':	4215 obs. of  5 variables:
  ..$ CHROMOSOME: chr [1:4215] "chr1" "chr1" "chr1" "chr1" ...
  ..$ POSITION  : int [1:4215] 4481813 4482616 4483161 4484382 4487057 5010631 5907041 6349089 9285059 9783132 ...
  ..$ END       : int [1:4215] 4481819 4482624 4483172 4484401 4487063 5010636 5907051 6349106 9285084 9783173 ...
  ..$ SCORE     : num [1:4215] 7.47 15.54 7.77 6.58 11.96 ...
  ..$ LENGTH    : num [1:4215] 7 9 12 20 7 6 11 18 26 42 ...
}
\details{
Chen X, Xu H, Yuan P, Fang F et al. Integration of external signaling pathways with the core transcriptional network in embryonic stem cells. Cell 2008 Jun 13;133(6):1106-17.
}
\source{
http://www.eh3.uc.edu/GenomicsPortals/showProp_tl.jsp?exp=GSE11431&db=Mm8Gff
}
\references{
Chen X, Xu H, Yuan P, Fang F et al. Integration of external signaling pathways with the core transcriptional network in embryonic stem cells. Cell 2008 Jun 13;133(6):1106-17. 
}
\examples{
data(mesFactors)
## maybe str(mesFactors) ; plot(mesFactors) ...
}
\keyword{datasets}
