% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_cells.R
\name{rank_cells}
\alias{rank_cells}
\title{Get the Rank of the Mean logcount for each Gene in each Cell}
\usage{
rank_cells(sce, group_col = "cellType")
}
\arguments{
\item{sce}{SingleCellExperiment Object with logcount assay}

\item{group_col}{name of the column in the colData of sce that defines the group of interest}
}
\value{

}
\description{
Get the Rank of the Mean logcount for each Gene in each Cell
}
\examples{
rank_cells(sce_zero_test, group_col = "group")
}
