% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_group.R
\name{rank_group}
\alias{rank_group}
\title{Get the Rank of the Mean logcount for each Gene in each Group}
\usage{
rank_group(sce, group_col = "cellType")
}
\arguments{
\item{sce}{SingleCellExperiment Object with logcount assay}

\item{group_col}{name of the column in the colData of sce that defines the group of interest}
}
\value{
Named list of ranks for each gene
}
\description{
Get the Rank of the Mean logcount for each Gene in each Group
}
\examples{
## TODO add comments
rank_group(sce_zero_test, group_col = "group")
}
