% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_invariance.R
\name{rank_invariance2}
\alias{rank_invariance2}
\title{Calculate the Rank Invariance of Each Gene from SCE object w/ groups}
\usage{
rank_invariance2(sce, group_col = "cellType")
}
\arguments{
\item{sce}{SingleCellExperiment Object with logcount assay}

\item{group_col}{name of the column in the colData of sce that defines the group of interest}
}
\value{
rank of invariance for each gene
}
\description{
Calculate the Rank Invariance of Each Gene from SCE object w/ groups
}
\examples{
rank_invariance2(sce_zero_test, group_col = "group")
}
\seealso{
Other invariance functions: 
\code{\link{rank_invariance}()}
}
\concept{invariance functions}
