% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_invariance.R
\name{rank_invariance_express}
\alias{rank_invariance_express}
\title{Calculate the Rank Invariance of Each Gene from SCE object w/ groups}
\usage{
rank_invariance_express(sce, group_col = "cellType", assay = "logcounts")
}
\arguments{
\item{sce}{SingleCellExperiment Object with logcount assay}

\item{group_col}{name of the column in the colData of sce that defines the group of interest}

\item{assay}{name of the assay in the sce object to use to rank}
}
\value{
rank of invariance for each gene
}
\description{
Calculate the Rank Invariance of Each Gene from SCE object w/ groups
}
\examples{
## Calulate RI for the sce object
rank_invariance_express(sce_zero_test, group_col = "group")
}
\seealso{
Other invariance functions: 
\code{\link{rank_invariance}()}
}
\concept{invariance functions}
