% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce_zero_test}
\alias{sce_zero_test}
\title{Test SummarizedExperiment data}
\format{
A \link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment-class}
object with 5 rows and 100 columns.
}
\usage{
sce_zero_test
}
\description{
A dataset with a range of Proportion Zeros across groups for different
genes which is used as a test and example dataset throughout this package.
}
\keyword{datasets}
