\name{treg-package}
\alias{treg-package}
\alias{treg}
\docType{package}
\title{
Signatures of transcription factor regulatory activity
}
\description{
The package provides tools for constructing gene-specific TF binding scores from ChIP-seq data (TREG binding scores) and for constructing TREG signatures of TF regulatory activity by integrating TREG binding scores with suitable matched differential gene expression profiles
}
\details{
\tabular{ll}{
Package: \tab treg\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-08-17\cr
License: \tab GPL (>= 2)\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Mario Medvedovic, Zhen Hu

Maintainer: Mario Medvedovic
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
