% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_prop_zero.R
\name{filter_prop_zero}
\alias{filter_prop_zero}
\title{Filter Genes for by Max Proportion Zero Among Groups}
\usage{
filter_prop_zero(prop_zero_df, cutoff = 0.9)
}
\arguments{
\item{prop_zero_df}{data.frame containing proportion of zero counts, genes as rows, groups as columns}

\item{cutoff}{Cutoff Value for Maximum Prop}
}
\value{
list of gene names that are under the cutoff
}
\description{
Filter Genes for by Max Proportion Zero Among Groups
}
\examples{
prop_zero <- get_prop_zero(sce_zero_test, group_col = "group")
filter_prop_zero(prop_zero, cutoff = 0.59)
}
