% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_prop_zero.R
\name{filter_prop_zero}
\alias{filter_prop_zero}
\title{Filter Genes for by Max Proportion Zero Among Groups}
\usage{
filter_prop_zero(prop_zero_df, cutoff = 0.9)
}
\arguments{
\item{prop_zero_df}{data.frame containing proportion of zero counts, genes as
rows, groups as columns.}

\item{cutoff}{Cutoff Value for Maximum Proportion Zero.}
}
\value{
A \code{character()} of gene names that are under the cutoff.
}
\description{
TODO add description.
}
\examples{
## Get proportion zero data.frame
prop_zero <- get_prop_zero(sce_zero_test, group_col = "group")

## Filter with max proportion zero cutoff = 0.59
filter_prop_zero(prop_zero, cutoff = 0.59)
}
