% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_zero.R
\name{get_prop_zero}
\alias{get_prop_zero}
\title{Get the Proption of Zero Counts for Each Gene in Each Group}
\usage{
get_prop_zero(sce, group_col = "cellType")
}
\arguments{
\item{sce}{\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment-class} Object}

\item{group_col}{name of the column in the colData of sce that defines the group of interest}
}
\value{
data.frame containing proportion of zero counts, genes as rows, groups as columns
}
\description{
Get the Proption of Zero Counts for Each Gene in Each Group
}
\examples{
rowSums(assays(sce_zero_test)$counts)
get_prop_zero(sce_zero_test)
get_prop_zero(sce_zero_test, group_col = "group")
}
