% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_zero.R
\name{get_prop_zero}
\alias{get_prop_zero}
\title{Get the Proportion of Zero Counts for Each Gene in Each Group}
\usage{
get_prop_zero(sce, group_col = "cellType")
}
\arguments{
\item{sce}{\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment-class} object}

\item{group_col}{name of the column in the
\link[SingleCellExperiment:SingleCellExperiment]{colData()} of \code{sce}
that defines the group of interest.}
}
\value{
A \code{data.frame()} containing proportion of zero counts, genes as rows,
groups as columns.
}
\description{
TODO add description.
}
\examples{
## Basic Proportion counts == 0
rowSums(assays(sce_zero_test)$counts == 0) / nrow(sce_zero_test)

## Get proportion by cell type
get_prop_zero(sce_zero_test)

## Get proportion by defined group
get_prop_zero(sce_zero_test, group_col = "group")
}
