% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_group.R
\name{rank_group}
\alias{rank_group}
\title{Get the Rank of the Mean expression for each Gene in each Group}
\usage{
rank_group(sce, group_col = "cellType", assay = "logcounts")
}
\arguments{
\item{sce}{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment} object with
the \code{assay} (defaults to \code{logcounts}).}

\item{group_col}{name of the column in the
\link[SummarizedExperiment:SummarizedExperiment-class]{colData()} of \code{sce}
that defines the group of interest.}

\item{assay}{A \code{character(1)} specifying the name of the
\link[SummarizedExperiment:SummarizedExperiment-class]{assay()} in the
\code{sce} object to use to rank expression values. Defaults to \code{logcounts} since
it typically contains the normalized expression values.}
}
\value{
Named \code{list()} of ranks for each gene
}
\description{
TODO description.
}
\examples{
## Rank the genes for each group defined by "group" column
rank_group(sce_zero_test, group_col = "group")
}
