% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_invariance.R
\name{rank_invariance}
\alias{rank_invariance}
\title{Calculate the Rank Invariance of Each Gene from Cell and Group Ranks}
\usage{
rank_invariance(group_rank, cell_rank)
}
\arguments{
\item{cell_rank}{}
}
\value{
rank of invariance for each gene
}
\description{
Calculate the Rank Invariance of Each Gene from Cell and Group Ranks
}
\examples{
group_rank_test <- rank_group(sce_zero_test, group_col = "group")
cell_rank_test <- rank_cells(sce_zero_test, group_col = "group")
rank_invar_test <- rank_invariance(group_rank_test, cell_rank_test)
}
\seealso{
Other invariance functions: 
\code{\link{rank_invariance2}()}
}
\concept{invariance functions}
