% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_invariance_express.R
\name{rank_invariance_express}
\alias{rank_invariance_express}
\title{Calculate the Rank Invariance of Each Gene from SCE object}
\usage{
rank_invariance_express(sce, group_col = "cellType", assay = "logcounts")
}
\arguments{
\item{sce}{\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment::SingleCellExperiment} object with
the \code{assay} (defaults to \code{logcounts}).}

\item{group_col}{name of the column in the
\link[SingleCellExperiment:SingleCellExperiment]{colData()} of \code{sce}
that defines the group of interest.}

\item{assay}{A \code{character(1)} specifying the name of the \code{assay()} in the
\code{sce} object to use to rank expression values. Defaults to \code{logcounts} since
it typically contains the normalized expression values.}
}
\value{
A \code{numeric()} with the rank of invariance for each gene. High values
represent TODO.
}
\description{
Description TODO.
}
\examples{
## Calculate RI for the sce object
rank_invariance_express(sce_zero_test, group_col = "group")
rank_invariance_express(sce_zero_test, group_col = "cellType")
}
\seealso{
Other invariance functions: 
\code{\link{rank_invariance}()}
}
\concept{invariance functions}
