% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce_zero_test}
\alias{sce_zero_test}
\title{Test SummarizedExperiment data}
\format{
A \link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment-class}
object with 5 rows and 100 columns.
}
\usage{
sce_zero_test
}
\description{
A simulated \link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment-class}
object representing the expression of 100 cells across 5 genes.
This object is used as an example dataset throughout \code{TREG}.
}
\details{
The \code{colData} of the object contains demo sample information about the cell,
that were designed to be used as \code{group_col}, (such as \code{cell_type}, \code{donor}, and  \code{region}).

The expression of the 5 genes were designed to show a range of
Proportion Zeros across groups for different genes.
The overall \code{prop_zero} for these theoretical genes are:
\itemize{
\item g100: \code{prop_zero = 1.00}
\item g50: \code{prop_zero = 0.50}
\item g0: \code{prop_zero = 0.00}
\item gOffOn: \code{prop_zero = 0.50}
\item gVar: \code{prop_zero = 0.54}
}
}
\keyword{datasets}
